Add Overridable Methods For Slot Amount Text
This is the cleanest solution to get slots to show >int max stack sizes.
These new methods are overridden in GT to use the AE stack size instead
of the vanilla stack size, which would be tricky to do in MUI itself
without strongly coupling the two.

This change will likely be pointless once MUI2 becomes standard, but for
now this is needed.